Add multiple extension support
Fixes #6

An alternative implementation to #17 that also supports linking between different file types.

When a relative url is discovered without a file extension then a set of HEAD requests are sent to different possible urls to try and discover one that works. In priority order are tried: the extensionless url, the url + the current pages extension, a list of extensions that can be configured by the user.

Configuration support has been added for Safari, should be relatively easy to do something similar for Chrome, I don't have it available to test though.

Test page for current pages extension (coffee):
  https://github.com/cabbagejs/cabbage/blob/master/lib/build-trees.coffee
Test page for across filetypes (js -> coffee):
  https://github.com/cabbagejs/cabbage/blob/master/index.js